refactor: moved next-auth.d.ts file
